more progress, backup. merged get_data into model.py
