+++ data/automate.py
@@ -1,44 +1,3 @@
-import sys
-
-import csv
-from time import sleep
-from fin_data import DailyTimeSeries
-from fracdiff import FractionalDifferentiation as fd
-
-import numpy as np
-import pandas as pd
-import os, sys
-
-from sklearn.preprocessing import MinMaxScaler
-
-from keras.models import Sequential
-from keras.layers import LSTM
-from keras.layers import Dense
-from keras.layers import RepeatVector
-from keras.layers import TimeDistributed
-
-from sklearn.model_selection import train_test_split
-from sklearn.metrics import r2_score
-
-from keras.preprocessing.sequence import TimeseriesGenerator
-from keras.callbacks import ModelCheckpoint
-
-
-class HiddenPrints:
-    def __enter__(self):
-        self._original_stdout = sys.stdout
-        sys.stdout = open(os.devnull, 'w')
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        sys.stdout.close()
-        sys.stdout = self._original_stdout
-
-
-pd.options.display.max_rows = 999
-pd.options.display.max_columns = 999
-
-import warnings
-warnings.simplefilter(action='ignore', category=FutureWarning)
