+++ data/model.py
@@ -1,17 +1,18 @@
-#import sys
-
-#sys.path.insert(0, '/home/joe/Documents/LambdaSchool/labs_ir/repos/Data-Science/data')
-
-from data_retrieval import Data
-
 import random
 import numpy as np
 import pandas as pd
 import os, sys
-import matplotlib.pyplot as plt
+import pickle
+import warnings
 
+from fin_data import DailyTimeSeries
+from fracdiff import FractionalDifferentiation as fd
+from keras_wrapper import KerasRegressorGenerator
+
+from time import sleep
+from sklearn.model_selection import GridSearchCV
 from sklearn.preprocessing import MinMaxScaler
-
+from sklearn.model_selection import train_test_split
 from keras.models import Sequential
 from keras.layers import LSTM
 from keras.layers import Dense
@@ -19,13 +20,6 @@ from keras.layers import RepeatVector
 from keras.layers import TimeDistributed
 from keras.layers import Dropout
 from keras import optimizers
-
-from sklearn.model_selection import train_test_split
-from sklearn.metrics import r2_score
-
-from keras.preprocessing.sequence import TimeseriesGenerator
-from keras.callbacks import ModelCheckpoint, EarlyStopping
-
 
 random.seed(42)
 
@@ -41,9 +35,6 @@ class HiddenPrints:
 
 pd.options.display.max_rows = 999
 pd.options.display.max_columns = 999
-
-import pickle
-import warnings
 
 warnings.simplefilter(action='ignore', category=FutureWarning)
 
@@ -63,7 +54,14 @@ class ModelMaker:
         self.technicals = technicals
         self.macros = macros
         self.shift = shift
+        self.X_train = None
+        self.X_test = None
+        self.X_val = None
         self.X_cols = None
+        self.y_train = None
+        self.y_test = None
+        self.y_val = None
+
 
     def create_model(self, lstm_layers=[361, 196, 81, 64],
                      dropouts=[0.2, 0.2, 0.2, 0.2],
@@ -81,15 +79,91 @@ class ModelMaker:
 
         sgd = optimizers.sgd(lr=l_rate, momentum=p, decay=decay)
 
-        model.compile(optimizer=sgd, loss=loss)
+        model.compile(optimizer=sgd, loss=loss, metrics=['accuracy'])
 
         return model
 
 #    def set_params(self, lstm_layers, dropouts, l_rate, momentum, decay, loss):
 
-    def fit_model(self, n_epochs, seq_lengths, batch_sizes):
-        dg = Data(self.symbol, self.technicals, self.macros, shift=1)
-        dg.get_data()
+    def fit_model(self, param_grid):
+        self.get_data()
+        krgmodel = KerasRegressorGenerator(build_fn=self.create_model,
+                                           val_data={'X':self.X_val, 'y':self.y_val})
 
-        for i in range(n_epochs):
-            train_data_generator, test_data_generator, val_data_generator = dg.create_gens(random.choice)
+        grid = GridSearchCV(estimator=krgmodel, param_grid=param_grid, n_jobs=-1)
+        grid_results=grid.fit(self.X_train, self.y_train)
+
+
+
+
+
+
+
+
+
+
+    def get_data(self, train_cut=.8, frac_diff_target=True):
+
+        dts = DailyTimeSeries(self.symbol)
+        df = dts.initiate()
+        sleep(5)
+
+        df = dts.add_fundamentals(df, self.fundamentals)
+        sleep(7)
+        df = dts.add_technicals(self.technicals, df)
+        sleep(15)
+        df = dts.add_macro(df, self.macros)
+
+        todrop = []
+
+        for col in df.columns:
+            if df[col].isnull().sum() / len(df) > .25:
+                print('{} is missing '.format(col), df[col].isnull().sum())
+                print('{} dropped'.format(col))
+                todrop.append(col)
+
+        df = df.drop(labels=todrop, axis=1)
+
+        if frac_diff_target:
+            if str(self.symbol + '_adjusted_close') in df.columns:
+                df['target'] = (fd.frac_diff_ffd(df[[str(self.symbol + '_adjusted_close')]],
+                                                 .35,
+                                                 .001).shift(-self.shift)[str(self.symbol + '_adjusted_close')])
+            else:
+                df['target'] = (fd.frac_diff_ffd(df[[str(self.symbol + '_close')]],
+                                                 .35,
+                                                 .001).shift(-self.shift)[str(self.symbol + '_close')])
+        else:
+            if str(self.symbol + '_adjusted_close') in df.columns:
+                df['target'] = df[str(self.symbol + '_adjusted_close')]
+            else:
+                df['target'] = df[str(self.symbol + '_close')]
+
+        df = df.dropna(axis=0)
+
+        print('nulls in target = ', df['target'].isnull().sum())
+        df = df.fillna(value=0)
+
+        X = df.drop(columns='target')
+        y = df[['target']].values
+
+        self.X_cols = X.columns
+
+        scaler = MinMaxScaler()
+        scaler.fit(X)
+
+        X_train, X_test, y_train, y_test = train_test_split(X, y,
+                                                            train_size=train_cut,
+                                                            shuffle=False)
+
+        X_test, X_val, y_test, y_val = train_test_split(X_test, y_test,
+                                                        train_size=.6,
+                                                        shuffle=False)
+        self.X_train = scaler.transform(X_train)
+        self.X_test = scaler.transform(X_test)
+        self.X_val = scaler.transform(X_val)
+        self.y_train = y_train
+        self.y_test = y_test
+        self.y_val = y_val
+
+        return None
